// src/models/enum/status.enum.ts
export enum StatusIdentifierEnum {
    MERGING = "merging",
    READY_FOR_MERGE = "ready_to_merge",
    MERGE_CONFLICTS = "merge_conflicts",
    UP_TO_DATE = "up_to_date"
}

export enum StatusNameEnum {
    MERGING = "Merging",
    READY_FOR_MERGE = "Ready to Merge",
    MERGE_CONFLICTS = "Merge Conflicts",
    UP_TO_DATE = "Up to Date"
}